refactoring for clarity and testing
